#13: Correct assembly name
