+++ RefleCS/RefleCS.TestKit/Nodes/BaseTypeBuilder.cs
@@ -1,6 +1,4 @@
 using RefleCS.Nodes;
-using ReflecCS.TestKit;
-using System;
 
 namespace RefleCS.TestKit.Nodes;
 public class BaseTypeBuilder : TestBuilderBase<BaseType>
