+++ RefleCS/RefleCS.TestKit/Nodes/ConstructorInitializerBuilder.cs
@@ -1,6 +1,5 @@
 using RefleCS.Enums;
 using RefleCS.Nodes;
-using ReflecCS.TestKit;
 
 namespace RefleCS.TestKit.Nodes;
 public class ConstructorInitializerBuilder : TestBuilderBase<ConstructorInitializer>
