+++ RefleCS/RefleCS.TestKit/Nodes/NamespaceBuilder.cs
@@ -1,5 +1,4 @@
 using RefleCS.Nodes;
-using ReflecCS.TestKit;
 
 namespace RefleCS.TestKit.Nodes;
 public class NamespaceBuilder : TestBuilderBase<Namespace>
