+++ RefleCS/RefleCS.TestKit/Nodes/PropertyBuilder.cs
@@ -1,8 +1,5 @@
 using RefleCS.Enums;
 using RefleCS.Nodes;
-using ReflecCS.TestKit;
-using System;
-using System.Collections.Generic;
 
 namespace RefleCS.TestKit.Nodes;
 public class PropertyBuilder : TestBuilderBase<Property>
