+++ RefleCS/RefleCS.TestKit/Nodes/RecordBuilder.cs
@@ -1,76 +1,86 @@
 using AutoFixture.Kernel;
-using ReflecCS.TestKit;
-using ReflecCS.TestKit.Common.Selectors;
 using RefleCS.Enums;
 using RefleCS.Nodes;
+using RefleCS.TestKit.Common.Selectors;
 
 namespace RefleCS.TestKit.Nodes;
 
-public class ClassBuilder : TestBuilderBase<Class>
+public class RecordBuilder : TestBuilderBase<Record>
 {
-    public ClassBuilder()
+    public RecordBuilder()
     {
-        Customize<Class>(c =>
+        Customize<Record>(c =>
             c.FromFactory(new MethodInvoker(new CtorSelectionQuery(typeof(IEnumerable<ClassModifier>)))));
     }
 
-    public ClassBuilder WithName(string name)
+    public RecordBuilder WithName(string name)
     {
         FillConstructorWith(nameof(name), name);
         return this;
     }
 
-    public ClassBuilder WithModifiers(IEnumerable<ClassModifier> modifiers)
+    public RecordBuilder WithModifiers(IEnumerable<ClassModifier> modifiers)
     {
         FillConstructorWith(nameof(modifiers), modifiers);
         return this;
     }
 
-    public ClassBuilder WithEmptyModifiers()
+    public RecordBuilder WithEmptyModifiers()
     {
         return WithModifiers(Enumerable.Empty<ClassModifier>());
     }
 
-    public ClassBuilder WithConstructors(IEnumerable<Constructor> constructors)
+    public RecordBuilder WithParameters(IEnumerable<Parameter> parameters)
+    {
+        FillConstructorWith(nameof(parameters), parameters);
+        return this;
+    }
+
+    public RecordBuilder WithEmptyParameters()
+    {
+        return WithParameters(Enumerable.Empty<Parameter>());
+    }
+
+    public RecordBuilder WithConstructors(IEnumerable<Constructor> constructors)
     {
         FillConstructorWith(nameof(constructors), constructors);
         return this;
     }
 
-    public ClassBuilder WithEmptyConstructors()
+    public RecordBuilder WithEmptyConstructors()
     {
         return WithConstructors(Enumerable.Empty<Constructor>());
     }
 
-    public ClassBuilder WithProperties(IEnumerable<Property> properties)
+    public RecordBuilder WithProperties(IEnumerable<Property> properties)
     {
         FillConstructorWith(nameof(properties), properties);
         return this;
     }
 
-    public ClassBuilder WithEmptyProperties()
+    public RecordBuilder WithEmptyProperties()
     {
         return WithProperties(Enumerable.Empty<Property>());
     }
 
-    public ClassBuilder WithMethods(IEnumerable<Method> methods)
+    public RecordBuilder WithMethods(IEnumerable<Method> methods)
     {
         FillConstructorWith(nameof(methods), methods);
         return this;
     }
 
-    public ClassBuilder WithEmptyMethods()
+    public RecordBuilder WithEmptyMethods()
     {
         return WithMethods(Enumerable.Empty<Method>());
     }
 
-    public ClassBuilder WithBaseTypes(IEnumerable<BaseType> baseTypes)
+    public RecordBuilder WithBaseTypes(IEnumerable<BaseType> baseTypes)
     {
         FillConstructorWith(nameof(baseTypes), baseTypes);
         return this;
     }
 
-    public ClassBuilder WithEmptyBaseTypes()
+    public RecordBuilder WithEmptyBaseTypes()
     {
         return WithBaseTypes(Enumerable.Empty<BaseType>());
     }
