+++ RefleCS/RefleCS.Tests/Nodes/ClassTests.cs
@@ -1,8 +1,8 @@
 ﻿using FluentAssertions;
-using ReflecCS.TestKit;
-using ReflecCS.TestKit.Common.Customizations;
 using RefleCS.Enums;
 using RefleCS.Nodes;
+using RefleCS.TestKit;
+using RefleCS.TestKit.Common.Customizations;
 using RefleCS.TestKit.Nodes;
 using RefleCS.TestTools.Exceptions;
