+++ RefleCS/RefleCS.Tests/Nodes/MethodTests.cs
@@ -1,9 +1,8 @@
 ﻿using FluentAssertions;
-using ReflecCS.TestKit;
-using ReflecCS.TestKit.Common.Customizations;
-using ReflecCS.TestKit.Nodes;
 using RefleCS.Enums;
 using RefleCS.Nodes;
+using RefleCS.TestKit;
+using RefleCS.TestKit.Common.Customizations;
 using RefleCS.TestKit.Nodes;
 using RefleCS.TestTools.Exceptions;
