+++ RefleCS/RefleCS.Tests/Nodes/NamespaceTests.cs
@@ -1,6 +1,6 @@
 ﻿using FluentAssertions;
-using ReflecCS.TestKit;
 using RefleCS.Nodes;
+using RefleCS.TestKit;
 using RefleCS.TestTools.Exceptions;
 
 namespace RefleCS.Tests.Nodes;
