+++ RefleCS/RefleCS.Tests/RefleCS.Tests.csproj
@@ -28,7 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ReflecCS.TestKit\ReflecCS.TestKit.csproj" />
+    <ProjectReference Include="..\ReflecCS.TestKit\RefleCS.TestKit.csproj" />
     <ProjectReference Include="..\RefleCS.TestTools\RefleCS.TestTools.csproj" />
     <ProjectReference Include="..\RefleCS\RefleCS.csproj" />
   </ItemGroup>
