+++ RefleCS/RefleCS.sln
@@ -12,9 +12,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RefleCS.TestTools", "RefleCS.TestTools\RefleCS.TestTools.csproj", "{5A96B93D-C564-4592-94C0-B2B62D028AB4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RefleCS.TestTools", "RefleCS.TestTools\RefleCS.TestTools.csproj", "{5A96B93D-C564-4592-94C0-B2B62D028AB4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReflecCS.TestKit", "ReflecCS.TestKit\ReflecCS.TestKit.csproj", "{A2BB1BB9-6451-4832-9E98-F30C4BC7CED0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RefleCS.TestKit", "RefleCS.TestKit\RefleCS.TestKit.csproj", "{A2BB1BB9-6451-4832-9E98-F30C4BC7CED0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
